Modify ConfirmApplyVC and add ConfirmType enum file
